Move emptyResponseCart to its own file
This is not a type, and should not be stored with types.
